为每个api添加数值；完善删除部门函数的逻辑

+++ offWorkSystem/src/main/java/org/off_work_system/web/DepartmentController.java
@@ -2,11 +2,15 @@ package org.off_work_system.web;
 
 import org.off_work_system.dto.ResultWrapper;
 import org.off_work_system.entity.Department;
+import org.off_work_system.entity.User;
+import org.off_work_system.enums.ResultStateEnum;
 import org.off_work_system.service.DepartmentService;
 import org.off_work_system.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 /**
  * Created by 张栋迪 on 2016/12/13.
@@ -27,6 +31,7 @@ public class DepartmentController {
             produces = {"application/json;charset=UTF-8"}
     )
     @ResponseBody
+    //添加新的部门，需要cookie中存储有管理员权限的用户
     public ResultWrapper<Department> addDepartment(
             @CookieValue(value = "userVerify", required = false) String userVerify,
             @ModelAttribute("departmentName") String departmentName,
@@ -54,12 +59,17 @@ public class DepartmentController {
             produces = {"application/json;charset=UTF-8"}
     )
     @ResponseBody
+    //传入用户id，删除该id的部门，要求该部门内无员工且是三个顶层部门以外的部门，需要cookie中存储有具有管理员权限的用户
     public ResultWrapper<Department> deleteDepartment(
             @CookieValue(value = "userVerify", required = false) String userVerify,
             @ModelAttribute("departmentId") int departmentId) {
         int permission = userService.verifyCookieOfAdmin(userVerify);
         if (permission != 200) {
             return new ResultWrapper<Department>(permission, null);
+        }
+        List<User> userList = userService.queryByDepartment(departmentId);
+        if (userList.size() != 0) {
+            return new ResultWrapper<Department>(605, null);
         }
         int result = departmentService.deleteDepartment(departmentId);
         return new ResultWrapper<Department>(result, null);
